added implementation for custom sound subscriber

+++ arduino/opencr_arduino/opencr/libraries/turtlebot3/include/turtlebot3/turtlebot3_sensor.h
@@ -81,7 +81,7 @@ class Turtlebot3Sensor
   // Sound
   void melody(uint16_t* note, uint8_t note_num, uint8_t* durations);
   void makeSound(uint8_t index);  
-  void makeCustomSound(uint8_t frequency, uint8_t duration);
+  void makeCustomSound(uint16_t frequency, uint16_t duration);
 
   // Bumper
   void initBumper(void);
